Move Lightbox component to kolibri/plugins/safe_html5_viewer/assets/src/views
