+++ kolibri/plugins/safe_html5_viewer/assets/src/views/Lightbox.vue
